Small Adjustment in Eval Engine (#126)
* adjust the distance function of CR and maxDist of KB & CR

* print all_cases

* skip the abstract input that contains no valid concrete values

* format

* sort benchmarks

* fix a benchmark

* use `long` instead of `int`

+++ xdsl_smt/eval_engine/src/Results.h
@@ -75,7 +75,8 @@ public:
     os << std::left << std::setw(20) << "num unsolved:" << unsolvedCases
        << "\n";
     os << std::left << std::setw(20)
-       << "base distance:" << baseDistance / maxDist(bw) << "\n";
+       << "base distance:" << static_cast<double>(baseDistance) / maxDist(bw)
+       << "\n";
     printMember(
         os, "num sound:", [](const Result &x) { return x.sound; },
         std::nullopt);
